finishing files & accounts

+++ src/Withdraw.java
@@ -1,8 +1,9 @@
-public class Withdraw extends Transaction{
-    public Withdraw(String date, int value, BankAccount bankAccount) {
+public class Withdraw extends Transaction {
+    public Withdraw(String date, double value, BankAccount bankAccount) {
         super(date, value, bankAccount);
         type = "Withdraw";
     }
+
     @Override
     public String toString() {
         return super.toString();
